requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
passlib
python-multipart
pycryptodomex
python-jose[cryptography]
bcrypt
mysqlclient
PyJWT
pytest
httpx
fastapi-limiter
pytest-mock